Add test for checkInstallNetworkManager()
