+++ install_test.go
@@ -17,6 +17,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -363,9 +364,37 @@ func connectTestClient(crtPath, keyPath string) (mqtt.Client, error) {
 	return mqttClient, nil
 }
 
+func isDockerInstalled() (bool, error) {
+	_, err := exec.LookPath("docker")
+	if err == nil {
+		return true, nil
+	}
+
+	return false, nil
+}
+
 func TestInstallDocker(t *testing.T) {
 	checkAndInstallDocker()
 	installed, err := isDockerInstalled()
 	assert.True(t, err == nil)
 	assert.True(t, installed)
+}
+
+func isNetManagerInstalled() bool {
+	cmd := exec.Command("bash", "-c", "dpkg --get-selections | grep network-manager")
+	out, _ := cmd.CombinedOutput()
+	return !strings.Contains(string(out), "deinstall") && len(out) != 0
+}
+
+func TestInstallNetworkManager(t *testing.T) {
+	assert.False(t, isNetManagerInstalled())
+	checkAndInstallNetworkManager()
+	defer func() {
+		c := exec.Command("bash", "-c", "apt-get remove -y network-manager")
+		_, err := c.CombinedOutput()
+		assert.True(t, err == nil)
+
+		assert.False(t, isNetManagerInstalled())
+	}()
+	assert.True(t, isNetManagerInstalled())
 }
